Deleted search.js
