+++ src/main/webapp/map.js
@@ -2,7 +2,9 @@ var map;
 var infoWindow;
 var shopInfo;
 var userPos;
-// require('dotenv').config();
+var geocoder;
+var markers = [];
+var coffeeShopInfo = [];
 
 /** Creates a map that shows all coffee shops around the user. */
 function createMap() {
@@ -10,7 +12,7 @@ function createMap() {
   var newyork = new google.maps.LatLng(40.7128, -74.0060);
   map = new google.maps.Map(
     document.getElementById('map'),
-    {center: newyork, zoom: 14});
+    {center: newyork, zoom: 13});
   
   // Infowindow for to handle error in get user location
   infoWindow = new google.maps.InfoWindow();
@@ -26,18 +28,7 @@ function createMap() {
           lng: position.coords.longitude
         };
         map.setCenter(userPos);
-
-        // Request nearby coffee shop info
-        // Unit of radius: metres. Maximum allowed is 5000 metres
-        var coffeshopRequest = {
-          location: userPos,
-          radius: '500',
-          query: 'coffee shop'
-        };
-        console.log(position.coords.latitude,position.coords.longitude);
-        console.log(google.maps.places);
-        service = new google.maps.places.PlacesService(map);
-        service.textSearch(coffeshopRequest, callback);
+        coffeeShopRequest(userPos);
       },
       function() {
         handleLocationError(true, infoWindow, map.getCenter());
@@ -49,21 +40,79 @@ function createMap() {
   }
 }
 
+// Request nearby coffee shop info
+// Unit of radius: metres. Maximum allowed is 50000 metres
+function coffeeShopRequest(userPos) {
+	var request = {
+    location: userPos,
+    radius: '300',
+    query: 'coffee shop'
+  };
+  service = new google.maps.places.PlacesService(map);
+  service.textSearch(request, callback);
+}
+
 function handleLocationError(browserHasGeolocation, infoWindow, userPos) {
   infoWindow.setPosition(userPos);
-  infoWindow.setContent(
-    browserHasGeolocation
-      ? "Error: The Geolocation service failed."
-      : "Error: Your browser doesn't support geolocation."
-  );
+  if (browserHasGeolocation) {
+  	infoWindow.setContent("The Geolocation service failed. Please enter a zip code to view nearby coffee shops.");
+  } else {
+  	infoWindow.setContent("Your browser doesn't support geolocation. Please enter a zip code to view nearby coffee shops.");
+  }
   infoWindow.open(map);
 }
 
-function callback(results, status) {
+// Call this wherever needed to actually handle the display
+function codeAddress() {
+	// Clear markers from previous search results
+	clearMarkers();
+	geocoder = new google.maps.Geocoder();
+	var zipCode = document.getElementById("zipcode").value;
+    geocoder.geocode( {
+      componentRestrictions: {
+        country: 'US',
+        postalCode: zipCode
+      }
+    }, function(results, status) {
+      if (status == google.maps.GeocoderStatus.OK) {
+        //Got result, center the map and put it out there
+        map.setCenter(results[0].geometry.location);
+        userPos = {
+          lat: results[0].geometry.location.lat(),
+          lng: results[0].geometry.location.lng()
+        };
+        coffeeShopRequest(userPos);
+      } else {
+        alert('Geocode was not successful for the following reason: ' + status);
+      }
+   });
+}
+  
+ function callback(results, status) {
   if (status == google.maps.places.PlacesServiceStatus.OK) {
-    for (var i = 0; i < results.length; i++) {
-      createMarker(results[i]);
+    if (document.URL.includes("store.html")) {
+      for (var i = 0; i < results.length; i++) {
+        createMarker(results[i]);
+      }
+    } else {
+      coffeeShopInfo = [];
+      for (var i = 0; i < results.length; i++) {
+        var distance = haversine_distance(results[i].geometry.location.lat(),
+          results[i].geometry.location.lng(), userPos.lat, userPos.lng);
+        // Create coffeeshop array that contains the name, address
+        // and distance from user
+        var coffeeShop = [];
+        coffeeShop.push(results[i].name);
+        coffeeShop.push(results[i].formatted_address);
+        coffeeShop.push(distance);
+        coffeeShopInfo.push(coffeeShop);
+        createMarker(results[i]);
+        coffeeShopInfo.sort( function(a, b) {
+          return (a[2] - b[2]);
+        });
+      }
     }
+    displayMarkers();
   }
 }
 
@@ -87,9 +136,68 @@ function createMarker(place) {
     localStorage.setItem("store", place.place_id);
     shopInfo.open(map, marker);
   });
+  markers.push(marker);
 }
 
-function getShopName() {
-  document.getElementById('title').innerHTML = localStorage.getItem("shopName");
-  document.getElementById('address').innerHTML = localStorage.getItem("address");
+function displayMarkers() {
+  for (var i = 0; i < markers.length; i++ ) {
+    markers[i].setMap(map);
+  }
+}
+
+function clearMarkers() {
+  for (var i = 0; i < markers.length; i++ ) {
+    markers[i].setMap(null);
+  }
+  markers = [];
+}
+
+function haversine_distance(lat1,lon1,lat2,lon2) {
+  var R = 6371; // Radius of the earth in km
+  var dLat = deg2rad(lat2-lat1);  // deg2rad below
+  var dLon = deg2rad(lon2-lon1); 
+  var a = 
+    Math.sin(dLat/2) * Math.sin(dLat/2) +
+    Math.cos(deg2rad(lat1)) * Math.cos(deg2rad(lat2)) * 
+    Math.sin(dLon/2) * Math.sin(dLon/2)
+    ; 
+  var c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1-a)); 
+  var d = R * c; // Distance in km
+  return d;
+}
+
+function deg2rad(deg) {
+  return deg * (Math.PI/180)
+}
+
+function calcRoute(userPos, coffeeShop) {
+	var directionsService = new google.maps.DirectionsService();
+  var request = {
+    origin: userPos,
+    destination: coffeeShop.geometry.location,
+    travelMode: 'WALKING'
+  };
+  directionsService.route(request, function(result, status) {
+    if (status == 'OK') {
+      coffeeShopInfo.push(coffeeShop.name, 
+      	coffeeShop.formatted_address, result.routes[0].legs[0].distance.text);
+    }
+  });
+}
+
+function displayRoute(userPos, coffeeShop) {
+  var directionsDisplay = new google.maps.DirectionsRenderer();
+  directionsDisplay.setMap(map);
+
+  var request = {
+    origin : userPos,
+    destination : coffeeShop,
+    travelMode : google.maps.TravelMode.WALKING
+  };
+  var directionsService = new google.maps.DirectionsService(); 
+  directionsService.route(request, function(response, status) {
+    if (status == google.maps.DirectionsStatus.OK) {
+      directionsDisplay.setDirections(response);
+    }
+  });
 }
